fix(cohorts): Use temporal for long running flag-copy task
